+++ src/Stryker.Core/Stryker.Core/Initialisation/InitialTestProcess.cs
@@ -9,7 +9,6 @@ namespace Stryker.Core.Initialisation
     public interface IInitialTestProcess
     {
         int InitialTest(ITestRunner testRunner);
-        TestCoverageInfos GetCoverage(ITestRunner testRunner);
         int TotalNumberOfTests { get; }
     }
 
@@ -31,6 +30,7 @@ namespace Stryker.Core.Initialisation
         public int InitialTest(ITestRunner testRunner)
         {
             var message = testRunner.DiscoverNumberOfTests() is var total && total == -1 ? "Unable to detect" : $"{total}";
+
             TotalNumberOfTests = total;
             _logger.LogInformation("Total number of tests found: {0}", message);
 
@@ -54,21 +54,5 @@ namespace Stryker.Core.Initialisation
             return (int)stopwatch.ElapsedMilliseconds;
         }
 
-        /// <summary>
-        /// Capture coverage informaiton
-        /// </summary>
-        /// <param name="testRunner"></param>
-        /// <returns></returns>
-        public TestCoverageInfos GetCoverage(ITestRunner testRunner)
-        {
-            var testResult = testRunner.CaptureCoverage();
-            if (!testResult.Success)
-            {
-                _logger.LogWarning("Test run with no active mutation failed. Stryker failed to correctly generate the mutated assembly. Please report this issue on github with a logfile of this run.");
-                throw new StrykerInputException("No active mutant testrun was not successful.", testResult.ResultMessage);
-            }
-
-            return testRunner.CoverageMutants;
-        }
     }
 }