+++ src/Stryker.Core/Stryker.Core/TestRunners/ITestRunner.cs
@@ -1,6 +1,6 @@
+using Stryker.Core.Mutants;
 using System;
 using System.Collections.Generic;
-using Stryker.Core.Mutants;
 
 namespace Stryker.Core.TestRunners
 {
@@ -9,7 +9,7 @@ namespace Stryker.Core.TestRunners
         TestRunResult RunAll(int? timeoutMS, IReadOnlyMutant activeMutant);
         int DiscoverNumberOfTests();
 
-        TestRunResult CaptureCoverage();
+        TestRunResult CaptureCoverage(bool cantUsePipe, bool cantUseUnloadAppDomain);
 
         TestCoverageInfos CoverageMutants { get; }