+++ src/Stryker.Core/Stryker.Core/TestRunners/TestCoverageInfos.cs
@@ -14,17 +14,6 @@ namespace Stryker.Core.TestRunners
         private readonly IList<TestDescription> _testsWithoutCoverageInfos = new List<TestDescription>();
         private static ILogger Logger { get; }
 
-        public IEnumerable<int> CoveredMutants
-        {
-            get
-            {
-                lock (_mutantToTests)
-                {
-                    return _mutantToTests.Keys;
-                }
-            }
-        }
-
         static TestCoverageInfos()
         {
             Logger = ApplicationLogging.LoggerFactory.CreateLogger<TestCoverageInfos>();
@@ -39,13 +28,8 @@ namespace Stryker.Core.TestRunners
         {
             lock (_mutantToTests)
             {
-                if (_mutantToTests.ContainsKey(mutant.Id))
-                {
-                    return _mutantToTests[mutant.Id].ToList();
-                }
-                
+                return _mutantToTests.ContainsKey(mutant.Id) ? _mutantToTests[mutant.Id].Union(_testsWithoutCoverageInfos).ToList() : _testsWithoutCoverageInfos;
             }
-            return null;
         }
 
         public void DeclareCoveredMutants(IEnumerable<int> list)
@@ -66,12 +50,12 @@ namespace Stryker.Core.TestRunners
                 if (captureCoverage == null)
                 {
                     // no coverage info available, we keep track of it
-                    Logger.LogDebug($"No covered mutants for {discoveredTest}.");
+                    Logger.LogDebug($"No mutant covered by {discoveredTest.Name}.");
                     _testsWithoutCoverageInfos.Add(discoveredTest);
                 }
                 else
                 {
-                    Logger.LogDebug($"Covered mutants for {discoveredTest} are: {string.Join(", ", captureCoverage)}.");
+                    Logger.LogDebug($"Mutants covered by {discoveredTest.Name} are: {string.Join(", ", captureCoverage)}.");
                     foreach (var id in captureCoverage)
                     {
                         if (_mutantToTests.ContainsKey(id))
@@ -83,6 +67,7 @@ namespace Stryker.Core.TestRunners
                             _mutantToTests.Add(id, new List<TestDescription>{discoveredTest});
                         }
                     }
+
                     if (staticMutants.Any())
                     {
                         Logger.LogDebug($"Those are being executed in a static constructor context: {string.Join(", ", staticMutants)}.");
@@ -95,14 +80,25 @@ namespace Stryker.Core.TestRunners
             }
         }
 
+        public bool IsCovered(int id)
+        {
+            return _testsWithoutCoverageInfos.Count > 0 || _mutantToTests.ContainsKey(id);
+        }
+
         public long UpdateMutants(IEnumerable<Mutant> mutants, int testsCount)
         {
             var avoidedTests = 0L;
             Logger.LogDebug("Optimize test runs according to coverage info.");
             var report = new StringBuilder();
             var mutantsToTest = mutants.Where(m => m.ResultStatus == MutantStatus.NotRun);
+
+            if (!mutantsToTest.Any())
+            {
+                return avoidedTests;
+            }
+
             var initialCount = mutantsToTest.Count();
-            var nonTested = mutants.Where(x => x.ResultStatus == MutantStatus.NotRun && !CoveredMutants.Contains(x.Id)).ToList();
+            var nonTested = mutants.Where(x => x.ResultStatus == MutantStatus.NotRun && !IsCovered(x.Id)).ToList();
             foreach (var mutant in nonTested)
             {
                 mutant.ResultStatus = MutantStatus.NoCoverage;
@@ -135,13 +131,14 @@ namespace Stryker.Core.TestRunners
             report.AppendJoin(',', nonTested.Select(x => x.Id));
             Logger.LogInformation(nonTested.Count == 0
                 ? "Congratulations, all mutants are covered by tests!"
-                : $"{nonTested.Count} mutants are not reached by any tests and will survive! (Marked as {MutantStatus.Survived}).");
+                : $"{nonTested.Count} mutant{(nonTested.Count>1 ? "s are" : " is")} not reached by any test and will survive! (Marked as '{MutantStatus.NoCoverage}').");
 
             var theoricalTestCount = (long)(testsCount) * initialCount;
-            Logger.LogInformation($"Coverage analysis eliminated {1.0*avoidedTests/theoricalTestCount:P} of tests (i.e. {avoidedTests} tests out of {theoricalTestCount}).");
+            var eliminatedTestPercentage = 1.0 * avoidedTests / theoricalTestCount;
+
+            Logger.LogInformation($"Coverage analysis eliminated {eliminatedTestPercentage:P} of tests (i.e. {avoidedTests} tests out of {theoricalTestCount}).");
 
             return avoidedTests;
         }
-
     }
 }