+++ src/Stryker.Core/CHANGELOG.md
@@ -1,3 +1,69 @@
+# [0.15.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.14.3...stryker@0.15.0) (2019-12-20)
+
+
+### Bug Fixes
+
+* **mutating:** Don't negate IsPatternExpression and properly rollback Accessor methods ([#846](https://github.com/stryker-mutator/stryker-net/issues/846)) ([69d7a71](https://github.com/stryker-mutator/stryker-net/commit/69d7a7106d6f5d54ba6e983ef3c58d30523f9578)), closes [#820](https://github.com/stryker-mutator/stryker-net/issues/820)
+
+
+### Features
+
+* **coverage analysis:** Implement defensive heuristics to cover when incorrect number of dynamic test cases are reported by testframework ([#852](https://github.com/stryker-mutator/stryker-net/issues/852)) ([470b08e](https://github.com/stryker-mutator/stryker-net/commit/470b08ee26afe3acbc2b96c91749b4245a6b13f6)), closes [#820](https://github.com/stryker-mutator/stryker-net/issues/820) [#843](https://github.com/stryker-mutator/stryker-net/issues/843) [#843](https://github.com/stryker-mutator/stryker-net/issues/843)
+* **Dashboard reporter:** Add dashboard reporter ([#849](https://github.com/stryker-mutator/stryker-net/issues/849)) ([7764472](https://github.com/stryker-mutator/stryker-net/commit/7764472fccc14759ee1c12499ba5aeccc0ba450a))
+
+
+### Reverts
+
+* Revert "Publish" ([8e2897f](https://github.com/stryker-mutator/stryker-net/commit/8e2897fb533ada246b56a9b5f50dd1d174c0d5cc))
+
+
+
+## [0.14.3](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.14.2...0.14.3) (2019-12-05)
+
+
+### Bug Fixes
+
+* **compilation:** Compile with AssemblyIdentityComparer compilation option ([#828](https://github.com/stryker-mutator/stryker-net/issues/828)) ([6eb3c16](https://github.com/stryker-mutator/stryker-net/commit/6eb3c163a098299dbe0e052d3d1dd970a298948d))
+* **targetframeworkversion extraction:** Correctly extract aggregated framework versions ([#827](https://github.com/stryker-mutator/stryker-net/issues/827)) ([85bdc99](https://github.com/stryker-mutator/stryker-net/commit/85bdc99b100666442eb46c9cf3db4b3473368db0))
+
+
+
+## [0.14.2](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.14.1...0.14.2) (2019-11-28)
+
+
+### Bug Fixes
+
+* **vstest:** Correctly determine targetframework for vstest configuration ([#822](https://github.com/stryker-mutator/stryker-net/issues/822)) ([cd2c7ff](https://github.com/stryker-mutator/stryker-net/commit/cd2c7ff1e8fde4fd3c3167d0c4332be01927d255))
+
+
+
+## [0.14.1](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.14.0...0.14.1) (2019-11-20)
+
+
+### Bug Fixes
+
+* **coverage analysis:** The name 'AppDomain' does not exist in the current context ([#789](https://github.com/stryker-mutator/stryker-net/issues/789)) ([d53686c](https://github.com/stryker-mutator/stryker-net/commit/d53686c57b24076630c927d99d72f1e7c7f41e11))
+
+
+
+# [0.14.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.13.0...0.14.0) (2019-11-15)
+
+
+### Bug Fixes
+
+* **linq mutator:** mutate linq query containing conditional access operator ([#761](https://github.com/stryker-mutator/stryker-net/issues/761)) ([d018b97](https://github.com/stryker-mutator/stryker-net/commit/d018b97358f0a9df8e879aa2c2296883e45a0da7)), closes [#760](https://github.com/stryker-mutator/stryker-net/issues/760)
+* **stryker-runner logging:** Use non-generic non-reflection ILogger.Log to log startup logging ([#744](https://github.com/stryker-mutator/stryker-net/issues/744)) ([64449c1](https://github.com/stryker-mutator/stryker-net/commit/64449c10adefc34bf43e219b5d3568a3e4502f53))
+
+
+### Features
+
+* **Bitwise mutations:** Add bitwise mutator ([#755](https://github.com/stryker-mutator/stryker-net/issues/755)) ([c684268](https://github.com/stryker-mutator/stryker-net/commit/c684268de4547b98a2d68381d1643ad50392d98c))
+* **Git diff:** Mutate only changed files based on git diff ([#708](https://github.com/stryker-mutator/stryker-net/issues/708)) ([34371c9](https://github.com/stryker-mutator/stryker-net/commit/34371c9bd47121f8c7458625718a229d8dfa0bee))
+* **negate mutator:** Add new negate mutations ([#773](https://github.com/stryker-mutator/stryker-net/issues/773)) ([8879654](https://github.com/stryker-mutator/stryker-net/commit/8879654f9170eaa276719ec3a3cefc6743009f2f))
+* **sharp8.0 nullable:** Compile with c# 8 nullable feature enabled ([#745](https://github.com/stryker-mutator/stryker-net/issues/745)) ([df11a97](https://github.com/stryker-mutator/stryker-net/commit/df11a978751752593c5b36bf4e3838913c59c03e))
+
+
+
 # [0.13.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@0.12.0...0.13.0) (2019-09-06)